apps: move device operation tests to a new file
